# Optimize JsonReader for the multi-segment case