@KrzysztofCwalina 2018-09-07
> The multi-segment case is now roughly 15-30% slower than the single-segment case (as long as the JSON payload is sufficiently large for the boundary crossing to get amortized). 

So it went down from 2x penalty (as you told me yesterday) to ~20% penalty for large segments?