@ahsonkhan 2018-09-07
> So it went down from 2x penalty (as you told me yesterday) to ~20% penalty for large segments?

Yes. The 2x penalty that I stated earlier was a ballpark estimate and was out-dated (didn't account for recent BufferReader and JsonReader changes). Accurately, the penalty was roughly **50-80%** (rather than 100%).

For example, in master, the Json4KB payload takes **8 us** for multi-segment (2 segments) vs **4.4 us** for single-segment. Now it takes **5.6 us**.