@KrzysztofCwalina 2018-09-07
OK, 20% penalty looks much more reasonable. Ship It! :-)